protect job creation, edit, and delete routes

+++ backend/job/views.py
@@ -2,9 +2,10 @@ from django.shortcuts import render, get_object_or_404
 from django.db.models import Avg, Min, Max, Count
 
 from rest_framework import status
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework.pagination import PageNumberPagination
+from rest_framework.permissions import IsAuthenticated
 
 from .serializers import JobSerializer
 from .models import Job
@@ -47,7 +48,9 @@ def getJob(request, pk):
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+    request.data['user'] = request.user
     data = request.data
 
     job = Job.objects.create(**data)
@@ -57,8 +60,12 @@ def newJob(request):
 
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def editJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.postedBy != request.user:
+        return Response({'message': 'You are not authorized to edit this job'}, status=status.HTTP_403_FORBIDDEN)
 
     job.title = request.data['title']
     job.description = request.data['description']
@@ -80,8 +87,12 @@ def editJob(request, pk):
 
 
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def deleteJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.postedBy != request.user:
+        return Response({'message': 'You are not authorized to delete this job'}, status=status.HTTP_403_FORBIDDEN)
 
     job.delete()
